Add grow families
I think I'll need to have two funcs; one to check if mothers become
pregnant and another to check if children can be born.

Tried to do the two at once to save on compute batches to/from the db,
but we're modifying what each other is looking for so the pagination is
off.